New certificate-update event. Refs #2969
The new certificate-update event always produces a new self-signed
certificate in the default location and launches a system-adjust
event.

The daily cronjob nethserver-check-certificate warns the admin if the
certificate is going to expire by sending an email.  If the system is
using the default self-signed certificate, a new certificate is
emitted automatically the day before the current certificate is
expired.